refactor: Inline streaming_router.py into chat_agent.py
Merge the single-consumer streaming_router module (124 lines) directly
into chat_agent.py. The routing logic (sync/async forced-RLM vs ReAct
dispatch) is now inlined in iter_chat_turn_stream/aiter_chat_turn_stream
methods plus a module-level _iter_forced_rlm_stream helper.

Part of Phase 3.1: agent helper consolidation.
